backend/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.8
mysql-connector==2.2.9
mysqlclient==1.4.6
PyYAML==3.13
pytest==5.3.1
pytest-flask-sqlalchemy==1.0.2
python-dotenv==0.13.0
